requirements.txt
pandas==2.0.3
numpy==1.23.5
scikit-surprise==1.1.3
optuna==3.3.0